CDP-1467
Modified DeleteProjects to use an HOC mutation instead of the Mutation component.
Added a delete function to DeleteProjects that deletes each project individually returning the results in a promise.
Modified TableActionsMenu to use HOC queries instead of Query component for more control. This also allows the calling of data.refetch
Added ActionResults components for handling success or error lists from a delete/unpublish action.
Extracted the error parsing function from ApolloError so that it can be used in the ActionResultError component.

Tests will need to be updated and have been left unmodified.